feat: initial debugger client implementation
The new debugger aims to be cross-platform easy to use
graphical application. Eventually it is going to provide
classic debugging techniques (like breakpoints), as well
as deep insight into tensors content.

To achieve this goal the debugger is implemented in C\#
with .NET Core. On Linux GUI is GTK3-based. macOS uses
Xamarin. And Windows uses WPF.