feat: introduce picomath kernel library (#71)
picomath is a portable library for linear algebra, machine
learning, equation solving. The initial release contains
an incomplete implementation of sgemm algorithm from BLAS,
and an OpenMP-based runtime. The library tries to follow
oneAPI specification as much as possible.

The SYCL implementation module is just a PoC. It hasn't
been tested, and it only presents to show that the kernel
code can be compiled for SYCL as well.